fixed procedural quotes issue for Update and Insert requests
